refactor(anthropic-direct): extract the tool-use round (6/7)
  loop/tool-round.ts     coordinator: assistant push, rollback, epilogue, cap
  loop/tool-dispatch.ts  build calls, start events, abort gate, dispatch
  loop/tool-results.ts   tool.output/tool.diff, block assembly, history commit

The rollback try/catch stays whole in tool-round.ts, spanning both
delegations. It has to: a throw between the assistant push and the
tool_result commit leaves history ending in an unmatched tool_use, which
400s every later API call. Splitting the push from its rollback across
modules is exactly the mistake that boundary exists to prevent, so the
invariant is now stated on the module rather than left implicit.

`iterations += 1` stays OUTSIDE the try, as before — a round that threw and
rolled back did not happen.

The render-channel isolation is preserved and documented: block assembly
destructures only content/isError/image, so ToolResult.render is
structurally unreachable at the push site rather than excluded by
convention.

loop.ts: 564 -> 267 lines. Every module in loop/ is now under the 350-line
house ceiling; the largest is stream-consumer.ts at 338.

Gates: pnpm lint clean; 526 tests pass across 51 files, unchanged.

Refs #360
