+++ src/agent/providers/anthropic-direct/loop.ts
@@ -25,37 +25,19 @@
  * @module agent/providers/anthropic-direct/loop
  */
 
-import type {
-  ContentBlockParam,
-  MessageParam,
-  ToolResultBlockParam,
-} from '@anthropic-ai/sdk/resources';
 import type { ProviderEvent } from '../../provider.js';
-import type { RunTurnInput, ToolCall, ToolResult } from './types.js';
+import type { RunTurnInput } from './types.js';
 import { toProviderUsage } from './types.js';
-import { translateMessageStream } from './translate.js';
-import { emitToolCall, emitSessionPhase } from '../../trace/emit.js';
-import { extractRawToolInput } from '../../facets/raw-input.js';
+import { emitSessionPhase } from '../../trace/emit.js';
 import { resolveTtfbTimeoutMs } from '../shared/first-byte-timeout.js';
 import { resolveStallTimeoutMs } from '../shared/stream-stall-timeout.js';
-import { summarizeToolInput } from '../shared/tool-input-summary.js';
-import {
-  TOOL_USE_LOOP_CAPPED,
-  WIND_DOWN_NOTE,
-  resolveMaxToolIterations,
-  shouldWindDown,
-} from '../shared/tool-loop-cap.js';
-import {
-  buildToolCallCompletedPayload,
-  buildToolCallStartedPayload,
-} from '../shared/tool-call-trace.js';
-import { DENIAL_BREAKER_FAILURE_CLASS } from '../../tools/denial-circuit-breaker.js';
-import { DenialCircuitBreakerError } from '../../../utils/errors.js';
+import { resolveMaxToolIterations } from '../shared/tool-loop-cap.js';
 import { OVERLOAD_EXHAUSTED, OVERLOAD_EXHAUSTED_NOTICE } from './overload-pause.js';
 import { OVERLOAD_MAX_RETRIES, RoundRetryBudget } from './loop/retry-budget.js';
 import { handleRoundRetry, type RoundRetryContext } from './loop/round-retry.js';
 import { openRound } from './loop/round-request.js';
 import { consumeRoundStream } from './loop/stream-consumer.js';
+import { runToolRound } from './loop/tool-round.js';
 import { emitNonToolUseTerminal } from './loop/turn-terminal.js';
 import { TurnAccumulator } from './loop/turn-accumulator.js';
 import { TurnTrace } from './loop/turn-trace.js';
@@ -273,290 +255,11 @@ export async function* runTurn(
       return;
     }
 
-    // stopReason === 'tool_use' — push the assistant turn into history,
-    // dispatch every tool_use block, then assemble the tool_result user turn.
-    //
-    // Rollback contract: capture the pre-push length so any throw between
-    // here and the final `input.messages.push(toolResultTurn)` below can
-    // splice the orphaned assistant message back out. Without this, an
-    // unexpected throw inside `executeBatch` / `execute` (one not absorbed
-    // into an `is_error: true` ToolResult) would leave history terminating
-    // in an unmatched `tool_use` — every subsequent API call would 400.
-    const messagesRollbackIdx = input.messages.length;
-    input.messages.push({
-      role: 'assistant',
-      content: turnResult.assistantBlocks,
-    });
-    try {
-
-    // Accumulate the cumulative tool-call tally BEFORE dispatch so the progress
-    // emit below reports calls-so-far including this round's batch. One round
-    // can carry N parallel tool_use blocks, so this advances by the block count,
-    // not by 1.
-    turn.toolCallCount += turnResult.toolUseBlocks.length;
-
-    // Build all tool calls and emit start events upfront.
-    const calls: ToolCall[] = [];
-    // Per-call start timestamps keyed by toolUseId so the completed
-    // trace event can carry an accurate `durationMs`. Lives within this
-    // loop iteration only — the next iteration starts fresh.
-    const startTimes = new Map<string, number>();
-    for (const block of turnResult.toolUseBlocks) {
-      calls.push({
-        id: block.id,
-        name: block.name,
-        input: block.input,
-        signal: input.signal,
-      });
-      const now = Date.now();
-      startTimes.set(block.id, now);
-      // Witness layer: tool_call.started fires BEFORE dispatch so even a
-      // crashing tool leaves evidence that it was attempted. Fire-and-
-      // forget — emitToolCall swallows writer errors internally.
-      void emitToolCall(
-        input.traceWriter,
-        buildToolCallStartedPayload({
-          toolUseId: block.id,
-          name: block.name,
-          input: block.input,
-          subagentId: input.subagentId,
-        }),
-      );
-      yield {
-        type: 'tool.use.start',
-        toolUseId: block.id,
-        toolName: block.name,
-        toolInput: summarizeToolInput(block.name, block.input),
-        toolInputRaw: extractRawToolInput(block.input),
-        sessionId: input.ctx.sessionId,
-      };
-    }
-
-    if (input.signal.aborted) {
-      const abortedResults: ToolResultBlockParam[] = calls.map((call) => ({
-        type: 'tool_result' as const,
-        tool_use_id: call.id,
-        content: 'Tool call aborted',
-        is_error: true,
-      }));
-      input.messages.push({ role: 'user', content: abortedResults as ContentBlockParam[] });
-      yield {
-        type: 'turn.completed',
-        usage: turn.terminalUsage(),
-        sessionId: input.ctx.sessionId,
-      };
-      return;
-    }
-
-    // Dispatch: batch (parallel for safe tools) or sequential fallback.
-    let results: ToolResult[];
-    if (input.toolDispatcher.executeBatch) {
-      try {
-        results = await input.toolDispatcher.executeBatch(calls);
-      } catch (err) {
-        results = calls.map(() => ({
-          content: `Tool batch execution failed: ${err instanceof Error ? err.message : String(err)}`,
-          isError: true as const,
-        }));
-      }
-    } else {
-      results = [];
-      for (const call of calls) {
-        if (input.signal.aborted) {
-          results.push({ content: 'Tool call aborted', isError: true });
-          continue;
-        }
-        try {
-          results.push(await input.toolDispatcher.execute(call));
-        } catch (err) {
-          const message = err instanceof Error ? err.message : String(err);
-          results.push({ content: `Tool execution threw: ${message}`, isError: true });
-        }
-      }
-    }
-
-    // Yield results and build tool_result blocks in original order.
-    const toolResultBlocks: ToolResultBlockParam[] = [];
-    for (let i = 0; i < calls.length; i++) {
-      const call = calls[i]!;
-      const result = results[i]!;
-
-      // Witness layer: tool_call.completed pairs with the .started event
-      // emitted above. `truncated` is now sourced from the handler's
-      // structured `ToolResult.truncated` flag — set by `handlers/bash.ts`
-      // and `handlers/grep.ts` whenever their byte cap is hit. The
-      // sentinel-substring fallback survives for third-party tool handlers
-      // that emit the `[output truncated …]` sentinel without setting the
-      // structured flag (back-compat). Fire-and-forget to keep the loop
-      // iteration cheap.
-      const startedAt = startTimes.get(call.id);
-      const durationMs = typeof startedAt === 'number' ? Date.now() - startedAt : 0;
-      const truncated = result.truncated === true || result.content.includes('[output truncated');
-      void emitToolCall(
-        input.traceWriter,
-        buildToolCallCompletedPayload({
-          toolUseId: call.id,
-          name: call.name,
-          result,
-          truncated,
-          durationMs,
-          subagentId: input.subagentId,
-        }),
-      );
-
-      yield {
-        type: 'tool.output',
-        toolUseId: call.id,
-        toolName: call.name,
-        content: result.content,
-        ...(result.isError === true ? { isError: true } : {}),
-        ...(truncated ? { truncated: true } : {}),
-        ...(result.incomplete === true ? { incomplete: true } : {}),
-        ...(result.incompleteReason ? { incompleteReason: result.incompleteReason } : {}),
-        ...(typeof result.batchIndex === 'number' && typeof result.batchSize === 'number'
-          ? { batchIndex: result.batchIndex, batchSize: result.batchSize }
-          : {}),
-        sessionId: input.ctx.sessionId,
-      };
-
-      // Sidecar render-only event for file-mutation tools. Travels on a
-      // separate event variant — the `toolResultBlocks.push()` call below
-      // cannot reference `result.render` (it's not in scope), so a future
-      // refactor cannot accidentally leak diff payloads into the model's
-      // `tool_result` content. This is the structural correctness invariant
-      // for the diff render channel.
-      if (result.render?.diff) {
-        yield {
-          type: 'tool.diff',
-          toolUseId: call.id,
-          diff: result.render.diff,
-          sessionId: input.ctx.sessionId,
-        };
-      }
-
-      // Destructure only the model-facing fields so `result.render` is
-      // structurally unreachable at this call site — not merely excluded by
-      // convention. This makes the isolation load-bearing rather than
-      // documentation-only. `image` is the ONE structured field that IS
-      // model-facing: when set it becomes an `image` content block alongside
-      // the text. `render` remains excluded.
-      const { content: resultContent, isError: resultIsError, image: resultImage } = result;
-      // When a tool returns an image (e.g. browser_screenshot), emit it as an
-      // image block followed by the text summary. The handler keeps the text
-      // non-empty so providers that drop the image still see useful context.
-      const toolResultContent: ToolResultBlockParam['content'] =
-        resultImage !== undefined
-          ? [
-              {
-                type: 'image',
-                source: {
-                  type: 'base64',
-                  media_type: resultImage.mediaType,
-                  data: resultImage.data,
-                },
-              },
-              ...(resultContent.length > 0
-                ? [{ type: 'text' as const, text: resultContent }]
-                : []),
-            ]
-          : resultContent;
-      toolResultBlocks.push({
-        type: 'tool_result',
-        tool_use_id: call.id,
-        content: toolResultContent,
-        ...(resultIsError === true ? { is_error: true } : {}),
-      });
-    }
-
-    const toolResultTurn: MessageParam = {
-      role: 'user',
-      content: toolResultBlocks as ContentBlockParam[],
-    };
-    input.messages.push(toolResultTurn);
-
-    // Denial circuit breaker (#546): the dispatcher tagged a result
-    // `denial-breaker` after a forked child hit N consecutive path-approval
-    // read denials with no progress. Surface it as a LOUD `error` event — never
-    // a silent partial — so the shared subagent handle rethrows it into a
-    // structured `buildResultFromError` failure the parent can act on (mirrors
-    // the TimeoutError teardown). History is already consistent here (assistant
-    // `tool_use` + matching `tool_result` blocks both pushed above), so
-    // returning is safe; the fork is done. A dispatcher throw would instead be
-    // swallowed by the executeBatch/execute catch above and the loop would keep
-    // spinning — which is the exact failure this breaker exists to prevent.
-    const denialTrip = results.find((r) => r.failureClass === DENIAL_BREAKER_FAILURE_CLASS);
-    if (denialTrip) {
-      yield { type: 'error', error: new DenialCircuitBreakerError(denialTrip.content) };
-      return;
-    }
-    } catch (err) {
-      // Rollback the orphaned assistant `tool_use` push above so the next
-      // turn's API call does not 400 with "tool_use ids were found without
-      // tool_result blocks immediately after". Any path that reached the
-      // matching `input.messages.push(toolResultTurn)` above (or the
-      // earlier aborted-results push at the signal-aborted gate) returned
-      // from inside the try with history already consistent, so they do
-      // not enter this catch. Re-throw so the outer query-level handler
-      // still surfaces the error.
-      input.messages.splice(messagesRollbackIdx);
-      throw err;
-    }
-
-    turn.iterations += 1;
-
-    const lastTool = turnResult.toolUseBlocks[turnResult.toolUseBlocks.length - 1];
-    // Semantic summary: name the tool AND its most informative argument
-    // (path / command / query via summarizeToolInput) so the progress banner
-    // carries real signal — `bash git show f7f0a37…` instead of the old
-    // `Iteration 11: used bash`, which conveyed nothing after round 2.
-    // `description` stays STABLE across ticks for this task id (the banner
-    // dedupes line 0 on commit); the per-tick signal lives in `summary`.
-    const lastToolHeadline = lastTool
-      ? `${lastTool.name}${summarizeToolInput(lastTool.name, lastTool.input)}`
-      : 'unknown';
-    yield {
-      type: 'progress',
-      progress: {
-        taskId: turn.taskId,
-        description: 'Working',
-        summary: `round ${turn.iterations}: ${lastToolHeadline}`,
-        lastToolName: lastTool?.name,
-        totalTokens: turn.usage.totalTokens ?? 0,
-        // Contract: `toolUses` is the cumulative COUNT OF TOOL CALLS so far in
-        // this turn (not the round number), so downstream formatToolCallStat
-        // renders "N tool calls" truthfully even when a round batched parallel
-        // calls. The `summary` above legitimately names the ROUND — leave it.
-        toolUses: turn.toolCallCount,
-        durationMs: Date.now() - turn.startedAt,
-      },
-      sessionId: input.ctx.sessionId,
-    };
-    if (turn.capReached) {
-      // The wind-down round (tools stripped) still came back as `tool_use` —
-      // only reachable if the model emits a tool call against an empty toolset.
-      // Honor the cap with a hard stop rather than looping unbounded.
-      yield {
-        type: 'turn.completed',
-        usage: turn.withDuration({ ...turn.usage, stopReason: TOOL_USE_LOOP_CAPPED }),
-        sessionId: input.ctx.sessionId,
-      };
-      return;
-    }
-    if (shouldWindDown(turn.iterations, maxIterations)) {
-      // Cap reached. Instead of cutting the turn off HERE — which ends it with
-      // no final assistant message and reads as a silent hang — run ONE more
-      // round with tools stripped (params build above) so the model can
-      // synthesize a final answer from what it gathered. Append a brief note to
-      // the tool_result turn just pushed so the model knows why its tools are
-      // gone. `turn.capReached` guarantees this fires at most once; the guard above
-      // hard-stops if the wind-down round pathologically emits another tool_use.
-      const lastMsg = input.messages[input.messages.length - 1];
-      if (lastMsg !== undefined && lastMsg.role === 'user' && Array.isArray(lastMsg.content)) {
-        lastMsg.content.push({ type: 'text', text: WIND_DOWN_NOTE });
-      }
-      turn.capReached = true;
-      continue;
-    }
+    // stopReason === 'tool_use' — dispatch the tools, commit the results, and
+    // decide whether the turn keeps going. The whole history mutation contract
+    // (assistant push, rollback on throw, tool_result commit) lives inside.
+    const round = yield* runToolRound(turnResult, input, turn, maxIterations);
+    if (round === 'terminated') return;
   }
   } finally {
     trace.finish(turn.elapsedMs());
